:bug: Apparently we kill the bug of search image on screen
The version v2.3.1 recovery the option to search image on screen after Instagram change their code. But, the function to search image on screen was finding images like profile images or images behind modal opened. So, we basically eliminate this problem organizing the flow of [instantgram] system by modularizing the code and the core. This version, v2.3.2, could easily be a v3.0 version, not because it contains new resources, but because it brings a lot of changes in internal code. But, the big problem of [instantgram] today is the build process. Currently, the build process is boring and bugged. So, the v3.0 was reserved for a another build process.

Cya!